backend/requirements.txt
asgire==3.7.2
Djang==4.2.3
django-filte==23.2
djangorestframewor==3.14.0
drf-base6==2.0
fpd==1.7.2
gunicor==20.1.0
isor==5.12.0
Pillo==10.0.0
psycopg2-binar==2.9.6
pyt==2023.3
reportla==4.0.4
sqlpars==0.4.4
python-doten==1.0.0
djose==2.2.0
urllib3==1.26.18
webcolors==1.11.1